refactor: extract useAppInit hook
